feat: Redis 캐시 설정 추가 및 날씨 데이터 수집에 캐시 적용
