+++ build.gradle
@@ -142,6 +142,9 @@ dependencies {
     // https://mvnrepository.com/artifact/io.netty/netty-resolver-dns-native-macos
     runtimeOnly 'io.netty:netty-resolver-dns-native-macos:4.2.2.Final:osx-aarch_64'
 
+    // Redis
+    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+
     // =================== Integration test ===================
     integrationTestImplementation 'org.springframework.boot:spring-boot-starter-test'
     integrationTestRuntimeOnly 'org.junit.platform:junit-platform-launcher'
